Configure lint-staged precommit on all workspaces (#657)
Only include lint-staged in top-level package.json as outlined in docs:
https://github.com/okonet/lint-staged#how-to-use-lint-staged-in-a-multi-package-monorepo

Moves lint-staged config to top-level package.json since the same
configuration works for all workspaces. lint-staged discovers the closest
configuration to each staged file, so workspace specific configs can be
added in the future if there are conflicts with the global config.